refactor: simplized import (#26)
