+++ backend/moneyFUS/models/__init__.py
@@ -1,3 +1,1 @@
-from .user.user import User as User, UserAuth as UserAuth
-from .asset.asset_log import AssetLog as AssetLog
-from .asset.usage_category import UsageCategory as UsageCategory
+from .user.user import User as User
