--- backend/moneyFUS/selializers/__init__.py
@@ -1,4 +1,0 @@
-from .user.user_serializer import (
-    UserSerializer as UserSerializer,
-    RegisterSerializer as RegisterSerializer,
-)
