+++ backend/moneyFUS/selializers/user/user_serializer.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 from django.contrib.auth.hashers import make_password
-from ...models import User, UserAuth
+from ...models.user.user import User, UserAuth
 
 
 class UserSerializer(serializers.ModelSerializer):
