+++ backend/moneyFUS/urls.py
@@ -1,11 +1,12 @@
 from django.urls import path
-from . import views
+from .views.user.index import UserView
+from .views.user.login import LoginView
 from .views.asset.index import AssetView
 from .views.asset.category import UsageCategoryView
 
 urlpatterns = [
-    path("user/login", views.user.login.LoginView.as_view()),
-    path("user", views.user.index.UserView.as_view()),
+    path("user/login", LoginView.as_view()),
+    path("user", UserView.as_view()),
     path("asset", AssetView.as_view()),
     path("asset/category", UsageCategoryView.as_view()),
 ]
